refactor(tasks): split the go_test file in several pieces
