+++ tasks/go.py
@@ -12,9 +12,9 @@ from invoke import task
 from invoke.exceptions import Exit
 
 from .build_tags import ALL_TAGS, UNIT_TEST_TAGS, get_default_build_tags
+from .libs.common.utils import get_build_flags, timed
 from .licenses import get_licenses_list
 from .modules import DEFAULT_MODULES, generate_dummy_package
-from .utils import get_build_flags, timed
 
 GOOS_MAPPING = {
     "win32": "windows",
