+++ tasks/go_test.py
@@ -5,7 +5,6 @@ High level testing tasks
 # Recent versions of Python should be able to use dict and list directly in type hints,
 # so we only need to check that we don't run this code with old Python versions.
 
-import abc
 import json
 import operator
 import os
@@ -13,7 +12,6 @@ import platform
 import re
 import sys
 from collections import defaultdict
-from contextlib import contextmanager
 from datetime import datetime
 from typing import Dict, List
 
@@ -25,18 +23,16 @@ from .build_tags import compute_build_tags_for_flavor
 from .cluster_agent import integration_tests as dca_integration_tests
 from .dogstatsd import integration_tests as dsd_integration_tests
 from .flavor import AgentFlavor
-from .go import run_golangci_lint
 from .libs.common.color import color_message
-from .libs.copyright import CopyrightLinter
+from .libs.common.utils import get_build_flags
 from .libs.datadog_api import create_count, send_metrics
-from .libs.junit_upload import add_flavor_to_junitxml, junit_upload_from_tgz, produce_junit_tar, repack_macos_junit_tar
+from .libs.junit_upload_core import add_flavor_to_junitxml, produce_junit_tar
 from .modules import DEFAULT_MODULES, GoModule
+from .test_core import ModuleTestResult, process_input_args, process_module_results, test_core
 from .trace_agent import integration_tests as trace_integration_tests
-from .utils import DEFAULT_BRANCH, get_build_flags
 
 PROFILE_COV = "coverage.out"
 GO_TEST_RESULT_TMP_JSON = 'module_test_output.json'
-UNIT_TEST_FILE_FORMAT = re.compile(r'[^a-zA-Z0-9_\-]')
 
 
 class TestProfiler:
@@ -69,265 +65,6 @@ def ensure_bytes(s):
         return s.encode('utf-8')
 
     return s
-
-
-@contextmanager
-def environ(env):
-    original_environ = os.environ.copy()
-    os.environ.update(env)
-    yield
-    for var in env:
-        if var in original_environ:
-            os.environ[var] = original_environ[var]
-        else:
-            os.environ.pop(var)
-
-
-TOOL_LIST = [
-    'github.com/frapposelli/wwhrd',
-    'github.com/go-enry/go-license-detector/v4/cmd/license-detector',
-    'github.com/golangci/golangci-lint/cmd/golangci-lint',
-    'github.com/goware/modvendor',
-    'github.com/stormcat24/protodep',
-    'gotest.tools/gotestsum',
-    'github.com/vektra/mockery/v2',
-]
-
-TOOL_LIST_PROTO = [
-    'github.com/favadi/protoc-go-inject-tag',
-    'github.com/grpc-ecosystem/grpc-gateway/protoc-gen-grpc-gateway',
-    'github.com/golang/protobuf/protoc-gen-go',
-    'github.com/golang/mock/mockgen',
-    'github.com/planetscale/vtprotobuf/cmd/protoc-gen-go-vtproto',
-    'github.com/tinylib/msgp',
-]
-
-TOOLS = {
-    'internal/tools': TOOL_LIST,
-    'internal/tools/proto': TOOL_LIST_PROTO,
-}
-
-
-@task
-def download_tools(ctx):
-    """Download all Go tools for testing."""
-    with environ({'GO111MODULE': 'on'}):
-        for path, _ in TOOLS.items():
-            with ctx.cd(path):
-                ctx.run("go mod download")
-
-
-@task
-def install_tools(ctx):
-    """Install all Go tools for testing."""
-    if os.path.isfile("go.work") or os.path.isfile("go.work.sum"):
-        # Someone reported issues with this command when using a go.work but other people
-        # use a go.work and don't have any issue, so the root cause is unclear.
-        # Printing a warning because it might help someone but not enforcing anything.
-
-        # The issue which was reported was that `go install` would fail with the following error:
-        ### no required module provides package <package>; to add it:
-        ### go get <package>
-        print(
-            color_message(
-                "WARNING: In case of issue, you might want to try disabling go workspaces by setting the environment variable GOWORK=off, or even deleting go.work and go.work.sum",
-                "orange",
-            )
-        )
-
-    with environ({'GO111MODULE': 'on'}):
-        for path, tools in TOOLS.items():
-            with ctx.cd(path):
-                for tool in tools:
-                    ctx.run(f"go install {tool}")
-
-
-@task
-def invoke_unit_tests(ctx):
-    """
-    Run the unit tests on the invoke tasks
-    """
-    for _, _, files in os.walk("tasks/unit-tests/"):
-        for file in files:
-            if file[-3:] == ".py" and file != "__init__.py" and not bool(UNIT_TEST_FILE_FORMAT.search(file[:-3])):
-                ctx.run(f"{sys.executable} -m tasks.unit-tests.{file[:-3]}", env={"GITLAB_TOKEN": "fake_token"})
-
-
-def test_core(
-    modules: List[GoModule],
-    flavor: AgentFlavor,
-    module_class: GoModule,
-    operation_name: str,
-    command,
-    skip_module_class: bool = False,
-    headless_mode: bool = False,
-):
-    """
-    Run the command function on each module of the modules list.
-    """
-    modules_results = []
-    if not headless_mode:
-        print(f"--- Flavor {flavor.name}: {operation_name}")
-    for module in modules:
-        module_result = None
-        if not skip_module_class:
-            module_result = module_class(path=module.full_path())
-        if not headless_mode:
-            skipped_header = "[Skipped]" if not module.condition() else ""
-            print(f"----- {skipped_header} Module '{module.full_path()}'")
-        if not module.condition():
-            continue
-
-        command(modules_results, module, module_result)
-    return modules_results
-
-
-class ModuleResult(abc.ABC):
-    def __init__(self, path):
-        # The full path of the module
-        self.path = path
-        # Whether the command failed for that module
-        self.failed = False
-        # String for representing the result type in printed output
-        self.result_type = "generic"
-
-    def failure_string(self, flavor):
-        return color_message(f"{self.result_type} for module {self.path} failed ({flavor.name} flavor)\n", "red")
-
-    @abc.abstractmethod
-    def get_failure(self, flavor):  # noqa: U100
-        """
-        Return a tuple with two elements:
-        * bool value - True if the result is failed, False otherwise
-        * str value - human-readable failure representation (if failed), empty string otherwise
-        """
-        pass
-
-
-class ModuleLintResult(ModuleResult):
-    def __init__(self, path):
-        super().__init__(path)
-        self.result_type = "Linters"
-        # Results of failed lint calls
-        self.lint_outputs = []
-
-    def get_failure(self, flavor):
-        failure_string = ""
-
-        if self.failed:
-            failure_string = self.failure_string(flavor)
-            failure_string += "Linter failures:\n"
-            for lint_output in self.lint_outputs:
-                if lint_output.exited != 0:
-                    failure_string = f"{failure_string}{lint_output.stdout}\n" if lint_output.stdout else failure_string
-                    failure_string = f"{failure_string}{lint_output.stderr}\n" if lint_output.stderr else failure_string
-
-        return self.failed, failure_string
-
-
-class ModuleTestResult(ModuleResult):
-    def __init__(self, path):
-        super().__init__(path)
-        self.result_type = "Tests"
-        # Path to the result.json file output by gotestsum (should always be present)
-        self.result_json_path = None
-        # Path to the junit file output by gotestsum (only present if specified in inv test)
-        self.junit_file_path = None
-
-    def get_failure(self, flavor):
-        failure_string = ""
-
-        if self.failed:
-            failure_string = self.failure_string(flavor)
-            failed_packages = set()
-            failed_tests = defaultdict(set)
-
-            # TODO(AP-1959): this logic is now repreated, with some variations, in three places:
-            # here, in system-probe.py, and in libs/pipeline_notifications.py
-            # We should have some common result.json parsing lib.
-            if self.result_json_path is not None and os.path.exists(self.result_json_path):
-                with open(self.result_json_path, encoding="utf-8") as tf:
-                    for line in tf:
-                        json_test = json.loads(line.strip())
-                        # This logic assumes that the lines in result.json are "in order", i.e. that retries
-                        # are logged after the initial test run.
-
-                        # The line is a "Package" line, but not a "Test" line.
-                        # We take these into account, because in some cases (panics, race conditions),
-                        # individual test failures are not reported, only a package-level failure is.
-                        if 'Package' in json_test and 'Test' not in json_test:
-                            package = json_test['Package']
-                            action = json_test["Action"]
-
-                            if action == "fail":
-                                failed_packages.add(package)
-                            elif action == "pass" and package in failed_tests:
-                                # The package was retried and fully succeeded, removing from the list of packages to report
-                                failed_packages.remove(package)
-
-                        # The line is a "Test" line.
-                        elif 'Package' in json_test and 'Test' in json_test:
-                            name = json_test['Test']
-                            package = json_test['Package']
-                            action = json_test["Action"]
-                            if action == "fail":
-                                failed_tests[package].add(name)
-                            elif action == "pass" and name in failed_tests.get(package, set()):
-                                # The test was retried and succeeded, removing from the list of tests to report
-                                failed_tests[package].remove(name)
-
-            if failed_packages:
-                failure_string += "Test failures:\n"
-                for package in sorted(failed_packages):
-                    tests = failed_tests.get(package, set())
-                    if not tests:
-                        failure_string += f"- {package} package failed due to panic / race condition\n"
-                    else:
-                        for name in sorted(tests):
-                            failure_string += f"- {package} {name}\n"
-            else:
-                failure_string += "The test command failed, but no test failures detected in the result json."
-
-        return self.failed, failure_string
-
-
-def lint_flavor(
-    ctx,
-    modules: List[GoModule],
-    flavor: AgentFlavor,
-    build_tags: List[str],
-    arch: str,
-    rtloader_root: bool,
-    concurrency: int,
-    timeout=None,
-    golangci_lint_kwargs: str = "",
-    headless_mode: bool = False,
-):
-    """
-    Runs linters for given flavor, build tags, and modules.
-    """
-
-    def command(module_results, module: GoModule, module_result):
-        with ctx.cd(module.full_path()):
-            lint_results = run_golangci_lint(
-                ctx,
-                module_path=module.path,
-                targets=module.lint_targets,
-                rtloader_root=rtloader_root,
-                build_tags=build_tags,
-                arch=arch,
-                concurrency=concurrency,
-                timeout=timeout,
-                golangci_lint_kwargs=golangci_lint_kwargs,
-                headless_mode=headless_mode,
-            )
-            for lint_result in lint_results:
-                module_result.lint_outputs.append(lint_result)
-                if lint_result.exited != 0:
-                    module_result.failed = True
-        module_results.append(module_result)
-
-    return test_core(modules, flavor, ModuleLintResult, "golangci_lint", command, headless_mode=headless_mode)
 
 
 def build_stdlib(
@@ -458,59 +195,6 @@ def codecov_flavor(
         ctx.run(f"codecov -f {path} -F {tag}", warn=True)
 
     return test_core(modules, flavor, None, "codecov upload", command, skip_module_class=True)
-
-
-def process_input_args(input_module, input_targets, input_flavors, headless_mode=False):
-    """
-    Takes the input module, targets and flavors arguments from inv test and inv codecov,
-    sets default values for them & casts them to the expected types.
-    """
-    if isinstance(input_module, str):
-        # when this function is called from the command line, targets are passed
-        # as comma separated tokens in a string
-        if isinstance(input_targets, str):
-            modules = [GoModule(input_module, targets=input_targets.split(','))]
-        else:
-            modules = [m for m in DEFAULT_MODULES.values() if m.path == input_module]
-    elif isinstance(input_targets, str):
-        modules = [GoModule(".", targets=input_targets.split(','))]
-    else:
-        if not headless_mode:
-            print("Using default modules and targets")
-        modules = DEFAULT_MODULES.values()
-
-    if not input_flavors:
-        flavors = [AgentFlavor.base]
-    else:
-        flavors = [AgentFlavor[f] for f in input_flavors]
-
-    return modules, flavors
-
-
-def process_module_results(module_results: Dict[str, Dict[str, List[ModuleResult]]]):
-    """
-    Expects results in the format:
-    {
-        "phase1": {
-            "flavor1": [module_result1, module_result2],
-            "flavor2": [module_result3, module_result4],
-        }
-    }
-
-    Prints failures, and returns False if at least one module failed in one phase.
-    """
-
-    success = True
-    for phase in module_results:
-        for flavor in module_results[phase]:
-            for module_result in module_results[phase][flavor]:
-                if module_result is not None:
-                    module_failed, failure_string = module_result.get_failure(flavor)
-                    success = success and (not module_failed)
-                    if module_failed:
-                        print(failure_string)
-
-    return success
 
 
 def deprecating_skip_linters_flag(skip_linters):
@@ -706,122 +390,6 @@ def test(
         raise Exit(code=1)
 
 
-def run_lint_go(
-    ctx,
-    module=None,
-    targets=None,
-    flavors=None,
-    build="lint",
-    build_tags=None,
-    build_include=None,
-    build_exclude=None,
-    rtloader_root=None,
-    arch="x64",
-    cpus=None,
-    timeout=None,
-    golangci_lint_kwargs="",
-    headless_mode=False,
-):
-    modules, flavors = process_input_args(module, targets, flavors, headless_mode)
-
-    linter_tags = {
-        f: build_tags
-        or compute_build_tags_for_flavor(
-            flavor=f, build=build, arch=arch, build_include=build_include, build_exclude=build_exclude
-        )
-        for f in flavors
-    }
-
-    modules_lint_results_per_flavor = {flavor: [] for flavor in flavors}
-
-    for flavor, build_tags in linter_tags.items():
-        modules_lint_results_per_flavor[flavor] = lint_flavor(
-            ctx,
-            modules=modules,
-            flavor=flavor,
-            build_tags=build_tags,
-            arch=arch,
-            rtloader_root=rtloader_root,
-            concurrency=cpus,
-            timeout=timeout,
-            golangci_lint_kwargs=golangci_lint_kwargs,
-            headless_mode=headless_mode,
-        )
-
-    return modules_lint_results_per_flavor
-
-
-@task(iterable=['flavors'])
-def lint_go(
-    ctx,
-    module=None,
-    targets=None,
-    flavors=None,
-    build="lint",
-    build_tags=None,
-    build_include=None,
-    build_exclude=None,
-    rtloader_root=None,
-    arch="x64",
-    cpus=None,
-    timeout: int = None,
-    golangci_lint_kwargs="",
-    headless_mode=False,
-):
-    """
-    Run go linters on the given module and targets.
-
-    A module should be provided as the path to one of the go modules in the repository.
-
-    Targets should be provided as a comma-separated list of relative paths within the given module.
-    If targets are provided but no module is set, the main module (".") is used.
-
-    If no module or target is set the tests are run against all modules and targets.
-
-    --timeout is the number of minutes after which the linter should time out.
-    --headless-mode allows you to output the result in a single json file.
-
-    Example invokation:
-        inv lint-go --targets=./pkg/collector/check,./pkg/aggregator
-        inv lint-go --module=.
-    """
-
-    # Format:
-    # {
-    #     "phase1": {
-    #         "flavor1": [module_result1, module_result2],
-    #         "flavor2": [module_result3, module_result4],
-    #     }
-    # }
-    modules_results_per_phase = defaultdict(dict)
-
-    modules_results_per_phase["lint"] = run_lint_go(
-        ctx=ctx,
-        module=module,
-        targets=targets,
-        flavors=flavors,
-        build=build,
-        build_tags=build_tags,
-        build_include=build_include,
-        build_exclude=build_exclude,
-        rtloader_root=rtloader_root,
-        arch=arch,
-        cpus=cpus,
-        timeout=timeout,
-        golangci_lint_kwargs=golangci_lint_kwargs,
-        headless_mode=headless_mode,
-    )
-
-    success = process_module_results(modules_results_per_phase)
-
-    if success:
-        if not headless_mode:
-            print(color_message("All linters passed", "green"))
-    else:
-        # Exit if any of the modules failed on any phase
-        raise Exit(code=1)
-
-
 @task(iterable=['flavors'])
 def codecov(
     ctx,
@@ -833,204 +401,6 @@ def codecov(
 
     for flavor in flavors:
         codecov_flavor(ctx, flavor, modules)
-
-
-@task
-def lint_teamassignment(_):
-    """
-    Make sure PRs are assigned a team label
-    """
-    branch = os.environ.get("BRANCH_NAME")
-    pr_url = os.environ.get("PR_ID")
-
-    if branch == DEFAULT_BRANCH:
-        print(f"Running on {DEFAULT_BRANCH}, skipping check for team assignment.")
-    elif pr_url:
-        import requests
-
-        pr_id = pr_url.rsplit('/')[-1]
-
-        res = requests.get(f"https://api.github.com/repos/DataDog/datadog-agent/issues/{pr_id}")
-        issue = res.json()
-
-        labels = {l['name'] for l in issue.get('labels', [])}
-        skip_qa_labels = ["qa/skip-qa", "qa/done", "qa/no-code-change"]
-        if any(skip_label in labels for skip_label in skip_qa_labels):
-            print("A label to skip QA is set -- no need for team assignment")
-            return
-
-        for label in labels:
-            if label.startswith('team/') and label != 'team/triage':
-                print(f"Team Assignment: {label}")
-                return
-
-        print(f"PR {pr_url} requires non-triage team assignment label (team/...); got labels:")
-        for label in labels:
-            print(f" {label}")
-        raise Exit(code=1)
-
-    # No PR is associated with this build: given that we have the "run only on PRs" setting activated,
-    # this can only happen when we're building on a tag. We don't need to check for a team assignment.
-    else:
-        print("PR not found, skipping check for team assignment.")
-
-
-@task
-def lint_skip_qa(_):
-    """
-    Ensure that when qa/skip-qa is used, we have one of [qa/done , qa/no-code-change]. Error if not valid.
-    """
-    branch = os.environ.get("BRANCH_NAME")
-    pr_url = os.environ.get("PR_ID")
-
-    if branch == DEFAULT_BRANCH:
-        print(f"Running on {DEFAULT_BRANCH}, skipping check for skip-qa label.")
-    elif pr_url:
-        import requests
-
-        pr_id = pr_url.rsplit('/')[-1]
-
-        res = requests.get(f"https://api.github.com/repos/DataDog/datadog-agent/issues/{pr_id}")
-        issue = res.json()
-
-        labels = {l['name'] for l in issue.get('labels', [])}
-        skip_qa = "qa/skip-qa"
-        new_qa_labels = ["qa/done", "qa/no-code-change"]
-        if skip_qa in labels and not any(skip_label in labels for skip_label in new_qa_labels):
-            print(
-                f"PR {pr_url} request to skip QA without justification. Requires an additional `qa/done` or `qa/no-code-change`."
-            )
-            raise Exit(code=1)
-        return
-    # No PR is associated with this build: given that we have the "run only on PRs" setting activated,
-    # this can only happen when we're building on a tag. We don't need to check for skip-qa.
-    else:
-        print("PR not found, skipping check for skip-qa.")
-
-
-@task
-def lint_milestone(_):
-    """
-    Make sure PRs are assigned a milestone
-    """
-    branch = os.environ.get("BRANCH_NAME")
-    pr_url = os.environ.get("PR_ID")
-
-    if branch == DEFAULT_BRANCH:
-        print(f"Running on {DEFAULT_BRANCH}, skipping check for milestone.")
-    elif pr_url:
-        import requests
-
-        pr_id = pr_url.rsplit('/')[-1]
-
-        res = requests.get(f"https://api.github.com/repos/DataDog/datadog-agent/issues/{pr_id}")
-        pr = res.json()
-        if pr.get("milestone"):
-            title = pr['milestone'].get('title', 'NO_TITLE')
-            print(f"Milestone: {title}")
-            if title != "Triage":
-                return
-
-        print(f"PR {pr_url} requires a non-Triage milestone.")
-        raise Exit(code=1)
-
-    # No PR is associated with this build: given that we have the "run only on PRs" setting activated,
-    # this can only happen when we're building on a tag. We don't need to check for a milestone.
-    else:
-        print("PR not found, skipping check for milestone.")
-
-
-@task
-def lint_releasenote(ctx):
-    """
-    Lint release notes with Reno
-    """
-
-    branch = os.environ.get("CIRCLE_BRANCH")
-    pr_url = os.environ.get("CIRCLE_PULL_REQUEST")
-
-    if branch == DEFAULT_BRANCH:
-        print(f"Running on {DEFAULT_BRANCH}, skipping release note check.")
-    # Check if a releasenote has been added/changed
-    elif pr_url:
-        import requests
-
-        pr_id = pr_url.rsplit('/')[-1]
-
-        # first check 'changelog/no-changelog' label
-        res = requests.get(f"https://api.github.com/repos/DataDog/datadog-agent/issues/{pr_id}")
-        issue = res.json()
-        if any([l['name'] == 'changelog/no-changelog' for l in issue.get('labels', {})]):
-            print("'changelog/no-changelog' label found on the PR: skipping linting")
-            return
-
-        # Then check that at least one note was touched by the PR
-        url = f"https://api.github.com/repos/DataDog/datadog-agent/pulls/{pr_id}/files"
-        # traverse paginated github response
-        while True:
-            res = requests.get(url)
-            files = res.json()
-            if any(
-                [
-                    f['filename'].startswith("releasenotes/notes/")
-                    or f['filename'].startswith("releasenotes-dca/notes/")
-                    or f['filename'].startswith("releasenotes-installscript/notes/")
-                    for f in files
-                ]
-            ):
-                break
-
-            if 'next' in res.links:
-                url = res.links['next']['url']
-            else:
-                print(
-                    "Error: No releasenote was found for this PR. Please add one using 'reno'"
-                    ", or apply the label 'changelog/no-changelog' to the PR."
-                )
-                raise Exit(code=1)
-    # No PR is associated with this build: given that we have the "run only on PRs" setting activated,
-    # this can only happen when we're building on a tag. We don't need to check for release notes.
-    else:
-        print("PR not found, skipping release note check.")
-
-    ctx.run("reno lint")
-
-
-@task
-def lint_filenames(ctx):
-    """
-    Scan files to ensure there are no filenames too long or containing illegal characters
-    """
-    files = ctx.run("git ls-files -z", hide=True).stdout.split("\0")
-    failure = False
-
-    if sys.platform == 'win32':
-        print("Running on windows, no need to check filenames for illegal characters")
-    else:
-        print("Checking filenames for illegal characters")
-        forbidden_chars = '<>:"\\|?*'
-        for file in files:
-            if any(char in file for char in forbidden_chars):
-                print(f"Error: Found illegal character in path {file}")
-                failure = True
-
-    print("Checking filename length")
-    # Approximated length of the prefix of the repo during the windows release build
-    prefix_length = 160
-    # Maximum length supported by the win32 API
-    max_length = 255
-    for file in files:
-        if (
-            not file.startswith(
-                ('test/kitchen/', 'tools/windows/DatadogAgentInstaller', 'test/workload-checks', 'test/regression')
-            )
-            and prefix_length + len(file) > max_length
-        ):
-            print(f"Error: path {file} is too long ({prefix_length + len(file) - max_length} characters too many)")
-            failure = True
-
-    if failure:
-        raise Exit(code=1)
 
 
 @task
@@ -1078,76 +448,6 @@ def e2e_tests(ctx, target="gitlab", agent_image="", dca_image="", argo_workflow=
             os.environ["ARGO_WORKFLOW"] = argo_workflow
 
     ctx.run(f"./test/e2e/scripts/setup-instance/00-entrypoint-{target}.sh")
-
-
-@task
-def lint_python(ctx):
-    """
-    Lints Python files.
-    See 'setup.cfg' and 'pyproject.toml' file for configuration.
-    If running locally, you probably want to use the pre-commit instead.
-    """
-
-    print(
-        f"""Remember to set up pre-commit to lint your files before committing:
-    https://github.com/DataDog/datadog-agent/blob/{DEFAULT_BRANCH}/docs/dev/agent_dev_env.md#pre-commit-hooks"""
-    )
-
-    ctx.run("flake8 .")
-    ctx.run("black --check --diff .")
-    ctx.run("isort --check-only --diff .")
-    ctx.run("vulture --ignore-decorators @task --ignore-names 'test_*,Test*' tasks")
-
-
-@task
-def lint_copyrights(_, fix=False, dry_run=False, debug=False):
-    """
-    Checks that all Go files contain the appropriate copyright header. If '--fix'
-    is provided as an option, it will try to fix problems as it finds them. If
-    '--dry_run' is provided when fixing, no changes to the files will be applied.
-    """
-
-    CopyrightLinter(debug=debug).assert_compliance(fix=fix, dry_run=dry_run)
-
-
-@task
-def install_shellcheck(ctx, version="0.8.0", destination="/usr/local/bin"):
-    """
-    Installs the requested version of shellcheck in the specified folder (by default /usr/local/bin).
-    Required to run the shellcheck pre-commit hook.
-    """
-
-    if sys.platform == 'win32':
-        print("shellcheck is not supported on Windows")
-        raise Exit(code=1)
-    if sys.platform.startswith('darwin'):
-        platform = "darwin"
-    if sys.platform.startswith('linux'):
-        platform = "linux"
-
-    ctx.run(
-        f"wget -qO- \"https://github.com/koalaman/shellcheck/releases/download/v{version}/shellcheck-v{version}.{platform}.x86_64.tar.xz\" | tar -xJv -C /tmp"
-    )
-    ctx.run(f"cp \"/tmp/shellcheck-v{version}/shellcheck\" {destination}")
-    ctx.run(f"rm -rf \"/tmp/shellcheck-v{version}\"")
-
-
-@task()
-def junit_upload(_, tgz_path):
-    """
-    Uploads JUnit XML files from an archive produced by the `test` task.
-    """
-
-    junit_upload_from_tgz(tgz_path)
-
-
-@task
-def junit_macos_repack(_, infile, outfile):
-    """
-    Repacks JUnit tgz file from macOS Github Action run, so it would
-    contain correct job name and job URL.
-    """
-    repack_macos_junit_tar(infile, outfile)
 
 
 @task
