+++ tasks/libs/github_actions_tools.py
@@ -9,9 +9,9 @@ from time import sleep
 
 from invoke.exceptions import Exit
 
-from ..utils import DEFAULT_BRANCH
 from .common.color import color_message
 from .common.github_api import GithubAPI
+from .common.utils import DEFAULT_BRANCH
 
 
 def trigger_macos_workflow(
