+++ tasks/libs/pipeline_tools.py
@@ -4,9 +4,9 @@ import platform
 import sys
 from time import sleep, time
 
-from ..utils import DEFAULT_BRANCH
 from .common.color import color_message
 from .common.user_interactions import yes_no_question
+from .common.utils import DEFAULT_BRANCH
 
 PIPELINE_FINISH_TIMEOUT_SEC = 3600 * 5
