+++ tasks/msi.py
@@ -12,7 +12,7 @@ from contextlib import contextmanager
 from invoke import task
 from invoke.exceptions import Exit, UnexpectedExit
 
-from tasks.utils import get_version, load_release_versions, timed
+from .libs.common.utils import get_version, load_release_versions, timed
 
 # Windows only import
 try:
