+++ tasks/release.py
@@ -17,10 +17,7 @@ from .libs.common.color import color_message
 from .libs.common.github_api import GithubAPI
 from .libs.common.gitlab import Gitlab, get_gitlab_token
 from .libs.common.user_interactions import yes_no_question
-from .libs.version import Version
-from .modules import DEFAULT_MODULES
-from .pipeline import edit_schedule, run
-from .utils import (
+from .libs.common.utils import (
     DEFAULT_BRANCH,
     GITHUB_REPO_NAME,
     check_clean_branch_state,
@@ -28,6 +25,9 @@ from .utils import (
     nightly_entry_for,
     release_entry_for,
 )
+from .libs.version import Version
+from .modules import DEFAULT_MODULES
+from .pipeline import edit_schedule, run
 
 # Generic version regex. Aims to match:
 # - X.Y.Z
