+++ tasks/systray.py
@@ -8,7 +8,7 @@ import sys
 
 from invoke import task
 
-from .utils import REPO_PATH, bin_name, get_version_ldflags, get_version_numeric_only
+from .libs.common.utils import REPO_PATH, bin_name, get_version_ldflags, get_version_numeric_only
 
 # constants
 BIN_PATH = os.path.join(".", "bin", "agent")
